Tagify library collects tags
Separate js file created for functions using tagify. Whitelists are loaded from database via AJAX calls. HTML file updated to include tags.js and necessary parts of tags.js are imported to scripts.js for event listeners.

**BUG**
If user edits tag to be something that is not in the whitelist, their original tag will remain in the system instead of being removed.
**BUG**